+++ src/Exceptions/CouldNotSendNotification.php
@@ -2,13 +2,11 @@
 
 namespace NotificationChannels\Twitter\Exceptions;
 
-class CouldNotSendNotification extends \Exception
+use Exception;
+
+class CouldNotSendNotification extends Exception
 {
-    /**
-     * @param $response
-     * @return CouldNotSendNotification
-     */
-    public static function serviceRespondsNotSuccessful($response)
+    public static function serviceRespondsNotSuccessful(mixed $response): CouldNotSendNotification
     {
         if (isset($response->error)) {
             return new static("Couldn't post notification. Response: ".$response->error);
@@ -19,23 +17,17 @@ class CouldNotSendNotification extends \Exception
         return new static("Couldn't post notification. Response: ".$responseBody);
     }
 
-    /**
-     * @param $response
-     * @return CouldNotSendNotification
-     */
-    public static function userWasNotFound($response)
+    public static function userWasNotFound(mixed $response): CouldNotSendNotification
     {
         $responseBody = print_r($response->errors[0]->message, true);
 
         return new static("Couldn't send direct message notification. Response: ".$responseBody);
     }
 
-    /**
-     * @param $exceededLength
-     * @return CouldNotSendNotification
-     */
-    public static function statusUpdateTooLong($exceededLength)
+    public static function statusUpdateTooLong(int $exceededLength): CouldNotSendNotification
     {
-        return new static("Couldn't post notification, because the status message was too long by ".$exceededLength.' character(s).');
+        return new static(
+            "Couldn't post notification, because the status message was too long by ${exceededLength} character(s)."
+        );
     }
 }