+++ src/TwitterChannel.php
@@ -3,37 +3,34 @@
 namespace NotificationChannels\Twitter;
 
 use Abraham\TwitterOAuth\TwitterOAuth;
-use Illuminate\Notifications\Notifiable;
 use Illuminate\Notifications\Notification;
 use NotificationChannels\Twitter\Exceptions\CouldNotSendNotification;
 
 class TwitterChannel
 {
-    /** @var TwitterOAuth */
-    protected $twitter;
-
-    /** @param TwitterOAuth $twitter */
-    public function __construct(TwitterOAuth $twitter)
+    public function __construct(protected TwitterOAuth $twitter)
     {
-        $this->twitter = $twitter;
     }
 
     /**
      * Send the given notification.
      *
-     * @param mixed $notifiable
-     * @param \Illuminate\Notifications\Notification $notification
+     * @param  mixed  $notifiable  Should be an object that uses the Illuminate\Notifications\Notifiable trait.
+     *
      * @throws CouldNotSendNotification
      */
-    public function send($notifiable, Notification $notification)
+    public function send($notifiable, Notification $notification): array|object
     {
         $this->changeTwitterSettingsIfNeeded($notifiable);
 
         $twitterMessage = $notification->toTwitter($notifiable);
         $twitterMessage = $this->addImagesIfGiven($twitterMessage);
 
-        $twitterApiResponse = $this->twitter->post($twitterMessage->getApiEndpoint(), $twitterMessage->getRequestBody($this->twitter),
-            $twitterMessage->isJsonRequest);
+        $twitterApiResponse = $this->twitter->post(
+            $twitterMessage->getApiEndpoint(),
+            $twitterMessage->getRequestBody(),
+            $twitterMessage->isJsonRequest,
+        );
 
         if ($this->twitter->getLastHttpCode() !== 200) {
             throw CouldNotSendNotification::serviceRespondsNotSuccessful($this->twitter->getLastBody());
@@ -45,23 +42,27 @@ class TwitterChannel
     /**
      * Use per user settings instead of default ones.
      *
-     * @param Notifiable $notifiable
+     * @param  object  $notifiable  Provide an object that uses the Illuminate\Notifications\Notifiable trait.
      */
-    private function changeTwitterSettingsIfNeeded($notifiable)
+    private function changeTwitterSettingsIfNeeded(object $notifiable)
     {
-        if (method_exists($notifiable, 'routeNotificationFor') && $twitterSettings = $notifiable->routeNotificationFor('twitter')) {
-            $this->twitter = new TwitterOAuth($twitterSettings[0], $twitterSettings[1], $twitterSettings[2],
-                $twitterSettings[3]);
+        if (
+            method_exists($notifiable, 'routeNotificationFor') &&
+            $twitterSettings = $notifiable->routeNotificationFor('twitter')
+        ) {
+            $this->twitter = new TwitterOAuth(
+                $twitterSettings[0],
+                $twitterSettings[1],
+                $twitterSettings[2],
+                $twitterSettings[3],
+            );
         }
     }
 
     /**
      * If it is a status update message and images are provided, add them.
-     *
-     * @param $twitterMessage
-     * @return mixed
      */
-    private function addImagesIfGiven($twitterMessage)
+    private function addImagesIfGiven(TwitterMessage $twitterMessage): object
     {
         if (is_a($twitterMessage, TwitterStatusUpdate::class) && $twitterMessage->getImages()) {
             $this->twitter->setTimeouts(10, 15);